+++ webapp/src/lib/marked/renderers/Blockquote.svelte
@@ -3,7 +3,7 @@
 
 	export let raw: string;
 
-	let isCallout: boolean = raw.split('\n')[0].match(/>\s?\[!(.+)\](\s.*|$)/) != null;
+	let isCallout: boolean = raw.split('\n')[0].match(/>\s?\[!(.+)\]([+-]?)(\s.*|$)/) != null;
 </script>
 
 {#if isCallout}