+++ webapp/src/test/markdown/callout.test.ts
@@ -2,19 +2,52 @@ import { render, screen } from '@testing-library/svelte';
 import { readMd } from './util';
 import MarkdownRenderer from '$lib/components/MarkdownRenderer.svelte';
 
-describe('Rendering callouts', async () => {
-	const plaintext = await readMd('callout.md');
+const testCases = [
+	{
+		title: 'Don!t forget to account for non-letters! //fsd \\n',
+		content: 'Sample text.',
+		markdown: `
+> [!Warning] Don!t forget to account for non-letters! //fsd \\n
+> Sample text.
+	`
+	},
+	{
+		title: 'Regular callout',
+		content: 'Sample text.',
+		markdown: `
+> [!NOTE] Regular callout
+> Sample text.
+	`
+	},
+	{
+		title: 'Collapsed callout',
+		content: 'Sample text.',
+		markdown: `
+> [!NOTE]- Collapsed callout
+> Sample text.
+	`
+	},
+	{
+		title: 'Uncollapsed callout',
+		content: 'Sample text.',
+		markdown: `
+> [!NOTE]+ Uncollapsed callout
+> Sample text.
+`
+	}
+];
 
+describe.each(testCases)('Rendering callouts', async (testCase) => {
 	it('Renders callout title correctly ', async () => {
-		render(MarkdownRenderer, { plaintext: plaintext });
-		const titleEl = await screen.findByText('Don!t forget to account for non-letters! //fsd \\n');
+		render(MarkdownRenderer, { plaintext: testCase.markdown });
+		const titleEl = await screen.findByText(testCase.title);
 		expect(titleEl).toBeInTheDocument();
 		expect(titleEl).toHaveClass('callout-title');
 	});
 
 	it('Renders callout content correctly ', async () => {
-		render(MarkdownRenderer, { plaintext: plaintext });
-		const contentEl = await screen.findByText('Sample text.');
+		render(MarkdownRenderer, { plaintext: testCase.markdown });
+		const contentEl = await screen.findByText(testCase.content);
 		expect(contentEl).toBeInTheDocument();
 		expect(contentEl.parentElement).toHaveClass('callout-content');
 	});